+++ QT_frontend/DLLRestAPI/engine_class.h
@@ -43,7 +43,7 @@ public:
     void login(QByteArray hash, long long card_num);
     void customer_info(int customer_id);
     void account_info(int account_id);
-    void card_status(int card_num);
+    void card_status(long long card_num);
     void transactions(int account_id, int index);
     void add_transaction(int account_id, double sum);
     void edit_balance(int account_id, double new_balance);
